+++ yaml/serializer.go
@@ -1,9 +1,9 @@
 package yaml
 
 import (
-	"os"
-
 	"io/ioutil"
+	"os"
+	"path"
 
 	"github.com/Scardiecat/svermaker"
 	"github.com/Scardiecat/svermaker/semver"
@@ -20,31 +20,34 @@ type Serializer struct {
 }
 
 type projectVersion struct {
-	current string
-	next    string
+	Current string
+	Next    string
 }
 
-func NewSerializer(path string) svermaker.Serializer {
+func NewSerializer(path string) *Serializer {
 	if path == "" {
-		path = "./"
+		path = "."
 	}
-	s := &Serializer{Path: path, Filename: "version.yml"}
+	s := &Serializer{Path: path, Filename: "/version.yml"}
 	s.projectVersionService.Serializer = s
 	return s
 }
 
 func (s *Serializer) Exists() bool {
-	return false
+	if _, err := os.Stat(path.Join(s.Path, s.Filename)); os.IsNotExist(err) {
+		return false
+	}
+	return true
 }
 
 func (s *Serializer) Serialize(p svermaker.ProjectVersion) error {
-	v := projectVersion{current: p.Current.String(), next: p.Next.String()}
+	v := projectVersion{Current: p.Current.String(), Next: p.Next.String()}
 	b, err := yaml.Marshal(&v)
 	if err != nil {
 		return err
 	}
 
-	err = ioutil.WriteFile(s.Path+s.Filename, b, 0755)
+	err = ioutil.WriteFile(path.Join(s.Path, s.Filename), b, 0755)
 	if err != nil {
 		return err
 	}
@@ -55,7 +58,7 @@ func (s *Serializer) Deserialize() (*svermaker.ProjectVersion, error) {
 	v := projectVersion{}
 	m := semver.Manipulator{}
 	projectVersion := svermaker.ProjectVersion{}
-	if file, err := os.Open(s.Path + s.Filename); err == nil {
+	if file, err := os.Open(path.Join(s.Path, s.Filename)); err == nil {
 
 		// make sure it gets closed
 		defer file.Close()
@@ -72,13 +75,13 @@ func (s *Serializer) Deserialize() (*svermaker.ProjectVersion, error) {
 	} else {
 		return nil, err
 	}
-	if current, err := m.Create(v.current); err == nil {
+	if current, err := m.Create(v.Current); err == nil {
 		projectVersion.Current = *current
 	} else {
 		return nil, err
 	}
 
-	if next, err := m.Create(v.next); err == nil {
+	if next, err := m.Create(v.Next); err == nil {
 		projectVersion.Next = *next
 	} else {
 		return nil, err